task: v0.0.1 - Mostly documentation
I just went through the process and reconfiguring another macbook, and
while I've had solutions in place for my virtual machine configuration
where I do most of my development, my core device was starkly lacking in
configuration management.  I looked around for different solutions for
this and decided to do a minimal implementation in the spirit of the
`mac-dev-playbook`. At this point the repo is totally untested, and
serves mostly as documentation, but at least I've documented most of the
required configuration.